Changes: - removed conditionally needing isDrivers in _layout.jsx - moved index.jsx, policy.jsx, sigin.jsx, and terms.jsx out of the scope of the navbar. Navbar doesn't appear on sign up or sign in pages. - replaced the "useNavigation" in signin.jsx with router replace for directing to profile page after sign-in - changed navbar icons for home and rider - minor cosmetic changes to toggle
