+++ desktop/project/app/terms.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
+
+export default function TermsScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Terms of Service</Text>
+      <Text style={styles.text}>This is a placeholder for Terms of Service.</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { flex: 1, padding: 24, backgroundColor: '#fff' },
+  title: { fontSize: 22, fontWeight: 'bold', marginBottom: 12 },
+  text: { fontSize: 16 },
+});
